Move client/{dist,__obj} under dist/client
