+++ .gitignore
@@ -1,9 +1,6 @@
 node_modules/
 npm-debug.log
 
-client/__obj/
-client/dist/
-
 dist/
 
 build/
