+++ client/index.html
@@ -285,9 +285,9 @@
     </div>
     </div>
 
-    <script defer src="dist/js/libs.min.js"></script>
-    <script defer src="dist/js/karen.templates.js"></script>
-    <script defer src="dist/js/karen.js"></script>
+    <script defer src="js/libs.min.js"></script>
+    <script defer src="js/karen.templates.js"></script>
+    <script defer src="js/karen.js"></script>
 
     </body>
 </html>
