+++ index.js
@@ -1,3 +1,3 @@
 #!/usr/bin/env node
 process.chdir(__dirname);
-var cli = require("./dist/command-line");
+var cli = require("./dist/server/command-line");
